twopass hal configuration
Allows multiple loadrt calls for same module

An earlier commit, aefefc1d9e2e6910, allows tcl scripts be used to configure
hal (jepler 8/19/2009 -- bravo!). A new directory is created, configs/simtcl,
for simulator examples that use .tcl files for hal configuration.

This patch uses the presence of an inifile entry [HAL]TWOPASS=anything to
direct a two-pass hal configuration using the files specified by
[HAL]HALFILE.  These HALFILE files can be existing .hal files or preferably
.tcl files which are very similar but allow use of programming constructions
like conditionals and loops.

  example:  HALFILE=twopass_demo.tcl
       or:  HALFILE=../sim/core_sim.hal

The principal script, twopass.tcl reads the HALFILE files in two passes.

In the first pass (pass0), no hal commands are executed but loadrt and loadusr
calls are consolidated.  This allows entry of loadrt calls for a given module
in multiple places using either the "count=" or "names=" specifiers (but not
simultaneously on a given module as they are mutually exclusive).

For compatibilty with existing .hal files, when .hal HALFILE files are
specified, they are first read and the [SECTION]ITEM specifiers are rewritten
as $::SECTION(ITEM) for tcl compatibility.  This step is not manifest to the
user.  The .hal files are also checked for hal commands which conflict in name
with tcl commands and warnings are issued to denote potential problems.

Since no hal commands are executed in pass0, the hal state cannot be
interrogated in any useful way.  At the conclusion of pass0, all the combined
loadrt and loadusr commands are executed.

In the second pass (pass1), the HAL[HALFILE] files are read again and all
commands except the loadrt and loadusr commands are executed (unless they are
for modules that were not detected in pass0).  A function (named
"::tp::passnumber") is available for use in tcl HALFILE files for debugging.

In the second pass, the addf functions are executed in the order of occurrence.
Since the order of the addf functions may be very important, the user must
manage this sequence.  A convenient method is to put all addf calls in one
place in the desired order.  An alternative method is to use the "position"
argument in the addf call.

Two demo/test cases based on sim/axis.ini and its three hal files
(core_sim.hal,axis_manualtoolchange.hal,simulated_home.hal)
are included:

twopass_compat.ini  uses unmodified hal files that are transcribed
                    transparently for tcl compatibility

twopass_demo.ini    uses a tcl configuration file with both loadrt "count="
                    and "names=" methods on different modules

Notes:
  1) Using the form "loadrt modname count=n" requires the user to manage
     the order and identity of the n modname items created.  Modules
     are created numbered 0 to n-1 and the user must keep track of the
     ordering.

     It is advantageous to use the form "loadrt modname names=namea,nameb ..."
     With this method, the user has explicit control of the component
     names and can use multiple loadrt lines even in different files.
     For example:

         filea
            loadrt flipflop names=ff1,ff2
         ...
         fileb
            loadrt flipflop names=ff9
         ...
         filec
            addf ff9 servo-thread
            addf ff2 servo-thread
            addf ff1 servo-thread

     With this method, configuration files can become more modular and
     readable while supporting an overall strategy for ordering the execution
     of the components.

  2) In executing [HAL]HALFILE tcl scripts, the tcl "puts" command will
     print during each pass.  The user can alter this behavior using procs:
        ::tp::passnumber   returns 0 | 1 for the current pass number
        ::tp::no_puts      disable the tcl "puts" command
        ::tp::alter_puts   alters  the tcl "puts" command so that it
                           reports <pass0> or <pass1>
        ::tp::restore_puts restores the tcl "puts" command