Add ability to load firmware from hostmot2 directories.
This gets is a start at having pncconf use hostmot2 firmware
from the hostmot2 directories. The biggest problem is the
assumtion pncconf made for the standard firmwares such as
the number of encoder pins for an encoder can be inconsistant
in some of the odd firmwares. I will have to blacklist some.