fix tests on ubuntu hardy
For whatever detailed reason, certain results that should mathematically
be exactly 0 end up having very tiny negative values on Ubuntu 8.04/i386
systems--for example, a result of -2.9e-17 instead of 0 in the test
g10-l1-l10.

In most cases, a result where the difference is not in the first 4
decimal places doesn't change the printed test results.  However, in
the specific instance of tiny negative value or negative zero instead of
a positive zero or tiny positive value, this can cause the test to fail
even though the software is working as intended.

These changes rewrite the string -0.0000 whereever it appears into 0.0000
so that this specific class of result difference does not lead to failed
tests.  Where -0.0000 was produced on Ubuntu 10.04/i386 machines, the
expected results have been changed accordingly.