build: fix warning during build
the latencyplot script needs to be rebuilt when the other .infiles
are rebuilt.

commit 2303fbc9 made this happen, but only for rt systems.  While I
understand the reasoning that latencyplot is not useful for sim
systems, this Makefile construction won't cause it not to be built
when initially configuring a sim system (that's done by the
configure script) , so we need to rebuild it from 'make' too.

Commit d003e440 made this happen for all systems, but didn't remove
the if-not-sim block, leading to the warning message on rt builds
only.