Disallow changing the active system with cutter comp on
While cutter comp is on, a nominal position and a compensated
position are both maintained.  If we allow changing the current
coordinate system, we would have to adjust both of those, and
probably also the endpoint in the cutter comp motion queue as well.
Especially in the stable branch, it's best to just disallow it.

Note that changing tool and G92 offsets were already disallowed.