comp: fix a typo in generated code
The comp tool accepts "option userinit yes", which lets the .comp author
provide a function to be executed before rtapi_app_main() and hal_init().
No .comp in 2.4, 2.5, or master currently uses this.

The documentation for comp specifies that this user-supplied function
be named userinit(), and the C code emitted by the comp tool calls the
function named userinit().

However, a typo in the comp tool causes it to emit a declaration for
a function named user_init() instead.  This commit fixes the emitted
declaration to name the function userinit().

Thanks to Carsten Presser for the bug report.