Allow Axisui Feedrate Override to display 9999%
When verifying g-code in an EMC2 simulation, it is sometimes very
convenient to run with an .ini file where all the velocities and
accelerations have been turned up to unreasonably high values,
just so you can run through the simulation faster.

For example, I'm working in a simulation now where the velocities
and accels have been turned up to about 20x actual values, and
the feedrate override max (slider) can go up to 1500% (15x normal).

This works just fine in EMC2 now, but the displayed value next to
the slider is limited to 3 digits, so when the slider exceeds
999%, it rolls over to 1000%, which displays as 000%. So when I
move my slider all the way up it only shows 500% instead of 1500%.

I noticed in 2.4.6 that the +1000% shows as 000%, but in 2.6.0pre
(current master) it shows as .000%. There seems to be a minor
artifact in the display that looks like a decimal point, but I
think it's the serif at the bottom of the one. Perhaps caused
by other display-related changes? No problem, just noting it.

I would recommend that this very minor fix be cherry-picked
from master to the more current revisions also, but I'll leave
that to the more experienced developers.

And if someone comes along that thinks 99x normal speed is
still too slow, then they can send in a patch to bump this up
again, from 4 digits to 5. Thanks for reading this far.

Signed-off-by: Kim Kirwan <Kim@KimKirwan.com>