iocontrol-v2: explicitly enable the start-change protocol by new -support-start-change option in ini/[EMCIO]EMCIO
reason:  this makes iocontrol-v2 a drop-in replacement for existing configurations. Before that, the hal config
needed to add a loop 'net start-change iocontrol.0.start-change  iocontrol.0.start-change-ack'
added to existing configurations so a an M6 would not block on waiting for start-change-ack.

Without the -support-start-change option, the start-change and start-change-ack pins are still created by iocontrol
but are inactive/ignored.

also, adapt demo ini file using iocontrol-v2 to use this option: axis-iocontrolv2-demo.ini and owordm6-ui.ini