ClassicLadder Modbus bug if more than 8 inputs read.
Les Newell posted to the emc-users mailing list:

ClassicLadder had a bug in it's Modbus implementation if you read
more than 8 inputs or coils. The first input bit is not read
correctly and ends up with strange values.
I don't know the correct way to submit a fix using Git but the fix
is simple. In protocol_modbus_master.c, a break; needs to be
inserted in line 321. Currently it falls through to the next case.

Bug fix submitted by Les Newell, thanks Les!

Git entry by Kim Kirwan, someone should check up on him though.
(I hope this is the fix that Les wanted?)

If someone more experienced will check it, maybe it should be
cherry-picked back to the 2.5.0 and 2.4.x versions?