interp/o_word: fix oword sub return & endsub errors when called from MDI and no file open
this removes the assumption that all o-word subs need to return/endsub to an
open .ngc file at the top level, even if in MDI mode (so the top level
ngc file isnt executing anyway).

This fix is a prerequisite for procedure-based toolchange to properly work
when no .ngc file is loaded.