let the user pass a branch to update-dch-from-git
Sane humans always work in the branch they currently have checked out
(ie, in the branch specified by .git/HEAD).

The buildbot, in contrast, is insane, and always works in the branch
"master", but with the working tree "git reset --hard" to the commit
it's building.

Thus for the buildbot's sake you need to pass the current branch in to
the tools that end up calling git describe.

This commit also makes debian/update-dch-from-git check for success or
failure of the get-version-from-git script.