fix a NULL pointer dereference
This commit fixes a bug that never bit.

The probe_new() function makes a window that lets the user select the
HAL object to display.  It takes a string argument intended to be the
title of the selection window.  The function intends to check the string
for NULL, and supply a default window title if the caller didn't provide
one, but it got the logic of the test reversed: if the caller provided
a title it always got replaced with the default, and if no title was
provided it would allow the NULL string through.

This bug never bit because probe_new() is only ever called by meter_new(),
which supplies a window title.