Docs: Added a date to manual front pages
I missed seeing the date on the front page of the manuals.
I don't claim to be an AsciiDoc expert, so if someone can
figure out a better way to do this, have at it.

The available choices seemed to be: (cut & pasted from)
http://www.methods.co.nz/asciidoc/chunked/ch27.html

{docdate}             document last modified date
{localdate}           the current date
{revdate}             document revision date (from document header)

So I went with {localdate} because I thought that was the one
with behavior most like it was before. Sorry if I picked the
wrong one. The title and date also seem to appear on each
page header. There are also {localtime}, {author}, and others
that might prove handy in the future.

I did this on both the English and French master txt documents.

Also updated the TODO list and did some minor edits.


Signed-off-by: Kim Kirwan <Kim@KimKirwan.com>