Docs: add dummy lyx files to Submakefile
I'm not sure that this is necessary, but it
might help, we'll see. I'm not sure exactly what
the docs scripting is doing. I'm still trying
things to see if I can get a good build here.

I'd guess that all the lyx files will be
eventually deleted in the 2.5 branch, and that's
fine, this is just a temporary test anyway.

Many of these files were just copied at random
from other files to fill space.


Signed-off-by: Kim Kirwan <Kim@KimKirwan.com>