Touchy: minor tweaking of screen layout for theme-change feature
Throughout touchy, buttons that are touching and stacked one above
the other, with one showing as "pressed in"/active and the others as
"out"/inactive, work as both radio buttons and status indicators.

These have lovingly been termed "butticators" by a user interface
expert who shall remain nameless.

The new Change Theme button was added to the hide/show pointer
butticator set, but it doesn't act in radiobutton fashion with
those other buttons.  To remove this confusion, I surrounded it
with whitespace so it looks like just a regular button (see also the
Override Limits button).  This also gives it slightly closer grouping
with the GTK Theme pulldown, with which it's conceptually linked.