ngcgui: support search path for filenames
For ngcgui tab pages embedded in axis gui, use a search path for the
filename items:
  [DISPLAY]NGCGUI_PREAMBLE, NGCGUI_SUBFILE, NGCGUI_POSTAMBLE, TTT_PREAMBLE

If a filename is specified with an initial "/" character, only the absolute
filename is considered.

When the file is specified as a relative filename (first character is not a "/"),
the search order is:

  [DISPLAY]PROGRAM_PREFIX
  [RS274NGC]SUBROUTINE_PATH

First file found in the search wins.

Thanks to Seb Kuzminsky for the suggestion.