Git: Add to .gitignore annoying .html files
For a long time now, "git status" has been showing a long list
of *.html files in /docs/src/<foo> that are auto-generated
for HTML docs. Let's see if we can shorten that list.
But apparently I have to commit the change to see if it works?
I'm only doing this on v2.5_branch, we can cherry-pick if it's good.


Signed-off-by: Kim Kirwan <Kim@KimKirwan.com>