Git: gitignore annoying .html, .lyx files
For a long time now, "git status" has been showing a long list
of leftover .lyx files we're not using anymore, and a long list
of *.html files that are auto-generated for HTML docs.
Let's see if we can shorten those lists.
I'm only doing this on v2.5_branch, we can cherry-pick or just
hand-edit if this is any good.


Signed-off-by: Kim Kirwan <Kim@KimKirwan.com>