Docs: Fix minor errors in Getting_Started (en,de,es,fr)
I saw that when a bold subject occurs, like this one...

Machine Name::

If you follow it with an index, like this...

Machine Name::(((Machine Name)))

It cancels the bold subject format and prints normally.
So I moved the index to the first line of text, like this:

Machine Name::
    (((Machine Name)))
    You should assign your machine a name, blah, blah, blah...

And I hope that fixes it. If this commit is immediately
followed by another one, it didn't, hopefully the 2nd one did.

I have to check all the other manuals for this problem,
but not tonight, tomorrow or Saturday, maybe.


Signed-off-by: Kim Kirwan <Kim@KimKirwan.com>