change debian version number format, again
This amends commit 7e8dbfd6f0c913bd277ee39edaa09e66892958a7, changing
the debian version number again.  7e8dbfd changed the debian version
number to be valid, but broke automatic upgrades: the new debian version
numbers had a + instead of a -, so the new numbers sorted as older than
the old numbers.

This changes the + to another ., which is valid and makes the version
number sort come out right.