twopass.tcl: support debug=dbg specifier
The pid component supports an option debug=dbg.
In the pid code, if the dbg >0, extra pins are exported
for all instances.

Here, twopass.tcl is modified so that occurrences of debug=dbg on
multiple loadrt lines are handled by ORing the values for
each line.  This maintains compatibility with the pid component
usage.

example:
   loadrt pid names=aa1 debug=0
   loadrt pid names=aa2 debug=1
   loadrt pid names=aa3 debug=0

will result in dbg = 0|1|0 = 1 and cause export of extra pins for all
pid instances.