make latencyplot fail gracefully if blt is missing
The latencyplot script checks for the availability of the BLT package
at startup.  If BLT is not available, it tries to suggest to the user
how to install it.

Before this commit, the script would fail to suggest the solution to the
user, instead crashing unhelpfully.  This commit moves the definition
of the "popup" function to before the first use, so it fails gracefully
with a helpful message instead of gracelessly with a stack trace.