only install latencyplot if it's going to work
latencyplot requires TCL 8.5 and the BLT package.  On Lucid this works
fine.  Hardy has TCL 8.5, but Hardy's BLT uses TCL 8.4, so it doesn't
work there.  This commit checks if latencyplot is going to work on the
target system, and skips it if it won't work.

This means we now build-depend on blt.