interp/T,M6,abort handlers: remove save/restore bracket of global state around handlers
the M6_COMMAND, M_COMMAND and ON_ABORT_COMMAND calls to Interp::execute()
were 'protected' by saving and restoring global state. This is is a bad
idea especially in the abort handler when e.g. settings are to be restored
to a known state - the 'restore global state' operation after the on_abort
NGC handler would revert such intended changes.