pncconf - upgrade firmware array to include smart serial
The firmware array needs space for smart serial info. Better to do this
before pncconf is released so it's possible to continue adding support
with out having to have many different size arrays, as that complicates
support of custom firmwares. This actually adds alot of code and glade
upgrades too. It can display comboboxes for smart serial but can't actually
do anything yet.

Signed-off-by: cmorley <chrisinnanaimo@hotmail.com>
Signed-off-by: Chris Morley <chrisinnanaimo@hotmail.com>
Signed-off-by: cmorley <chrisinnanaimo@hotmail.com>