pncconf -change hal HAL signalname enable to machine-is-enabled
motion.motion-enabled now will always have a signal named
machine-is-enabled connected to it. The mesa combo box
showed 'machine-is-enabled' but the HAL name was 'enable'
all the tests are looking for 'machine-is-enabled' so lets
just go with it. I actually had it this way once then switched
back to enable cause it caused errors when loading old configs.
now if pncconf doesn't recognise a signal it should just set it to
unused. By now pncconf has changed enough one should just
build a new config anyways.

Signed-off-by: cmorley <chrisinnanaimo@hotmail.com>