pncconf - fix missing custom.hal custom_postgui.hal files
These file were not always written or would be overwritten by
pncconf display options.
Now they are always written, if missing. They are for users.
pncconf options are written to other files. pncconf uses a 'call
file' to call pncconf options (if used)  then the user custom files.
While not the usual way things are done it is consistant and easier
to incorporate user customizations

Signed-off-by: Chris Morley <chrisinnanaimo@hotmail.com>
Signed-off-by: cmorley <chrisinnanaimo@hotmail.com>