pncconf - fix overwriting of call_list.hal & sticky sseial signals
If you choose existing custom program with pyvcp or gvcp
the call_list.hal would be written without calling the options
hal file. The user is probably relying on it be called. So
now as long as it is still there it would overwrite it when one
chooses existing.

Sserial signals selected from channels that were set to unused
afterwards were not cleared, so a signal name search would still
find them leading to connections to pins that were not really there.
Now they are cleared.
Also changed firmware numbering of sserial channels starting at 0
instead of 1. this is more consistent within pncconf and on the
7i44 the channels are numbered 0-7. sinse noone has the custom
firmware array to use sserial in pncconf yet its not a big deal.

Signed-off-by: cmorley <chrisinnanaimo@hotmail.com>