pncconf -fix gpio bug
sserial GPIO cannot be changed by user from input to output
as mainboard GPIO can. To designate this the component number
is set to 100 for sserial GPIO. previously if the compnum was 0
then it could be changed and anything higher could not. The
problem was when other components were deselected to cause them
to be GPIO, pncconf got confused because the compnum was still
coming from the original component, which often was higher then
0 but could actually be changed from input to ouput by the user.