pncconf -add basic support for the 5i25 board
add a 'not used' pintype so that the pins beyound 34 are not
shown ( the 5i25 only has 34 not 48 )
The extra pins are set to GPIOI with unused_input signal names
so that pncconf will ignore them yet I don't have to add code
to _every_ method that handles pintype.
since I don't have a 5i25 to test with the HAL code for loading
the card might not be right yet as the 5i25 doesn't load the firmware
each time EMC loads.That will be easy to fix later.
unfortunately hostmot2 repository doesn't have the 5i25 firmware XML
file, so users will have to use pncconf's back door method of adding
it.