interp/oword: missed a merge conflict
also, unsure why these fail during rebase:

Could not apply 8ca8d1c...
Don't assume that the firmware declares ioports first. There has been a tacit assumption in the sserial code that by the time it is called the ioports are all set up and ready to go. This is not necessarily a safe assumption. This patch defers sserial processing until all the other functions are finished. (checked with a special backwards bitfile made by Pete)

could not apply ec2a988...
A neater way to handle the possibility of ioports not being first on the list.