use the installed a2x instead of our own copy
Back in the beginning of 2011, we switched to asciidoc for our
documentation.  Part of that switch was using the a2x program to make
pdfs from the asciidoc sources.  Unfortunately the asciidoc in Ubuntu
Hardy is too old (older than 8.5) and lacks some of the features we want.

In the interest of expediency, a newer version of a2x was copied from
the asciidoc upstream into our repo, and we used that and moved on.

Some time later, Jeff Epler backported a modern asciidoc from Lucid to
Hardy, and published it in the linuxcnc.org debian archive:

    deb http://linuxcnc.org/hardy hardy base

Now that a proper packaged version of asciidoc's a2x is conveniently
available, there is no need for us to carry along our own copy any more.
So now, in the interest of hygiene, I'm removing our copy and pointing
the build system at the copy installed by debian package in /usr/bin.

Note that both our debian/control and our src/configure scripts check
for valid versions of asciidoc.  This commit also improves the configure
error a bit if the available asciidoc is too old.