interp/remap: introduce block.builtin_used flag
this enables detection wether a remap procedure referred to the code
being remapped (the "recursion case").

Rationale: testing this flag in the epilog permits 'doing the right thing
automatically' - if the code being remapped was referred to and hence
builtin semantics were executed, the epilog shouldnt do the same thing.

Howerver, if the remap procedure did NOT refer to the code being remapped,
then the epilog needs to take appropriate action.