teach configure to determine the python version
Michael Haberler's awesome g-code remapping patchbomb needs to link
a bunch of stuff against libpython, which requires versioned shared
libraries.

This commit teaches the src/configure script how to ask python what it's
version is, and constructs the library name from that version string.
It also switches a couple of Submakefiles to use this autoconfigured
library version instead of a hardcoded value (that worked on Lucid
but not on Hardy).  This commit makes it build on Hardy and Lucid and
hopefully any other distribution that includes python distutils.