Docs: Minor fix to userforeword_fr.txt
Apparently the abbreviation for Monsieur (M.), when it occurs at
the beginning of a line, is taken as some kind of numbered list?

Of the form:

J.
K.
L.
M.
etc.

So this is an attempt to fix that by getting rid of the M.
that appears at the start of a line.

Signed-off-by: Kim Kirwan <Kim@KimKirwan.com>