interp/M7x: dont fail quietly
this was too optimistic, and the underlying cause which led to ID: 3437928

This was triggered by an aborted G38.3 which left G38.3 as modal state.
the next oword call to a sub protected with M7x failed on return because
G38.3 cannot sensibly be 'restored to', which happened with the
manual-toolchange-with-tool-length-switch remap demo

Since the interpreter has no chance to do anything sensible when a G38
is aborted in a sub, the only proper resolution is to clear the modal
state in the on_abort handler (e.g. with a G80)