packaging: teach debian/configure about Precise Pangolin
debian/configure does a bunch of stuff, including modify the build
dependencies based on the distribution we're building on.  It uses
some fairly out-of-date defaults if it doesn't recognize the distro.
This commit teaches that part of debian/configure about Ubuntu Precise
Pangolin 12.04, so it can get the build dependencies it needs.