build system: make reasonable version numbers for unknown branches
The scripts/get-version-from-git script is used to produce a version
number based on things like the current branch and most recent tag signed
by the release managers.

It used to fail completely on branches other than 2.4, 2.5, and master.

This commit makes it do something hopefully reasonable on branches it
doesn't know about: it uses the version from the VERSION file, plus "~",
plus the branch name, plus another "~", plus the sha1 of the HEAD commit.