motion: remove useless call to ceil()
This call was useless, and reportedly causes some unspecified older
compilers to encounter internal compiler errors.  Others may warn
about the implicit conversion from int to double (as this makes the
code incompatible with pre-ANSI C compilers, something we all care
deeply about).

The original change may have intended to be something more like
    ceil((double)traj_period_nsec / nsec)
but since all we really care about is that the servo multiplier is
an exact integer, it's preferable to preserve the existing behavior
and just remove the useless function call.