experimental!  Add support for Sivar 7m43 board
Isssy's "Sivar 7m43" board is very similar to the Mesa 7i43, but has
more IO pins instead of the on-board LEDs.

This commit is a first try at adding support for this new board.

This commit needs to be tested before it goes into master, and the
manpage still needs to be updated.