packaging: fix dependencies for python-{gnome,glade}2
In Hardy, python-gnome2 Provides python2.5-gnome2 and python-glade2
Provides python2.5-glade2.

In Lucid, python-gnome2 Provides python2.6-gnome2 and python-glade2
Provides python2.6-glade2.

But in Precise, there are no such Provides.  Precise ships with python2.7
and python3, but some of the python module packages (specifically gnome2
and glade2) don't supply the Provides lines that Hardy and Lucid do.

This commit changes the emc2{,-sim} Depends to accept either
python@PYTHON_VERSION@-gnome2 or python-gnome2, and same for glade2.
Precise's python-gnome2 is for python2.7 (and same for glade2), so this
works well on all three supported distros.