gladevcp -expose more display options to hal_gremlin
can select such things as displaying the DRO in metric or imperial units.
If also checks the INI file, under [DISPLAY] for LATHE = 1 (or true or True).
This will then only allow P and Y views. The y view will be rotated to a more
common lathe layout and manual rotation will not be allowed. The DRO will
display radius and diameter for the X axis.