make comp emissions rename-friendly
This renames all variants of "emc2" to LinuxCNC in the MODULE_INFO
sections, the manpage headers, and command-line usage info.  Also links
userspace components with -llinuxcnchal (that's a mouthful!) instead of
the old library.

There is one place in this file where the old emc2 name survives: in the
paths to search for the Makefile.modinc file.  There will be a second
pass of this rename cleanup later...