support autoconf 2.68 (for Ubuntu 12.04)
Precise has autoconf 2.68 and gcc 4.6.2, Lucid has autoconf 2.64 and gcc
4.4.3, and Hardy has autoconf 2.61 and gcc 4.2.4.  All three versions
of gcc provide the sincos() function as a built-in.

Our configure.in script used to use AC_CHECK_FUNCS(sincos) to see if
sincos was available.  This worked with autoconf 2.61 and 2.64, but not
with 2.68.  autoconf 2.68 changed AC_CHECK_FUNCS() to "#undef sincos"
before trying to use it.  This fails with gcc builtin functions.

This commit changes the check for sincos() from AC_CHECK_FUNCS to
AC_TRY_LINK(), which works with all three of the versions of autoconf
that we support.