fix package relationships after rename
This commit makes the following things happen:

    * linuxcnc Conflicts with emc2 and emc2-sim

    * linuxcnc-sim Conflicts with emc2 and emc2-sim

    * linuxcnc-dev Conflicts with emc2-dev

    * linuxcnc Replaces emc2

    * linuxcnc-sim Replaces emc2-sim

    * linuxcnc-dev Replaces emc2-dev

This should make for smooth upgrades from the old emc2 packages to the
new linuxcnc packages.

This commit also reverts a mistake: the linuxcnc-doc-* packages still
need to Conflict with the old emc2 packages (which have the docs in the
main deb), not with the new linuxcnc packages (which have never had any
docs to conflict with).