attempt to work around an intermittent build error
When doing builds with "-j5", I have twice encountered a failure
like this:
    cp: cannot create regular file `../docs/html/common/images/naive-cam.png'cp: cannot create regular file `../docs/html/common/images/naive-cam.png': File exists
    : File exists
    mkdir -p ../docs/html/common
    cp: cannot create regular file `../docs/html/common/images/naive-cam.png': File exists
    make: *** [../docs/html/common/User_Concepts_de.html] Error 1
    make: *** Waiting for unfinished jobs....
    make: *** [../docs/html/common/User_Concepts_es.html] Error 1
    make: *** [../docs/html/common/User_Concepts_fr.html] Error 1
Tell cp to try copying harder.

I'm not 100% convinced this is the solution to the problem, because
(even though I saw it twice in one day yesterday!) I can't reproduce
the problem consistently enough to satisfy myself it no longer
happens after this change.