ngcgui: warn for absolute file spec conflict
Expected ini parameters for specifying files are:
         [RS274NGC]SUBROUTINE_PATH
     and
         [DISPLAY]NGCGUI_SUBFILE -- as a filename only

Unexpected usage: user specifies [DISPLAY]NGCGUI_SUBFILE as
                  an absolute filename

This fails, as expected, if absolute filename is not in search path.

Change is to warn if user specifies an absolute filename that doses not resolve
to the search path and simultaneously has a file by that name in the search
path.