live plotter: begin support of foam live plot
the foam live plot abuses the r[xyz] fields to store the uvw
coordinates; z and w are set by a new set_depth method call; the
live plot adds a new line when a certain distance has elapsed or a
corner is turned.

the "c" field is moved and the "c2" field is added so that
glDrawArrays can be used to draw the foam-type live plot.
glDrawArrays requires each vertex to have a color.  In non-foam
machines, the memory overhead of the unused "c2" field is 1/8 of the
total memory used for the live plot, or about 40kB for a 10000-point
live plot.