halcmd: Introduce and use A_TILDE
Some halcmd commands (save, source, loadusr) refer to filesystem paths.
Enable the use of ~ to refer to the home directory.

It may be necessary to offer a way to turn this off.  For instance,
directly running at the commandline of a traditional bourne-style shell:
   $ halcmd 'source ~/.hal'
should probably be permitted as a way to source the file ./~/.hal, since
the shell user has gone to the trouble of quoting ~ from the shell.

It's also not entirely clear whether halcmd_parse_cmd's contract with
its callers was to preserve the passed-in argument array; if so, a
temporary copy of the array will be required.  However, for all the
in-tree users this alternation of the argument array does not appear
to be problematic.