gladevcp/8.04 compat: apply Gtk bug workaround for gtkrc files only to GtkBuilder files
old (libglade) format files will not work properly with gtkrc
This is an 8.04 issue where you had to create glade UI's on 10.04 and save
them in libglade format, transfer them to 8.04, and execute them there
(editing gladevcp panels on 8.04 does not work)

Practically this only affects 2.5 gladevcp examples, most of which are
stored in libglade format so they be executed on 8.04. Anybody building a
new UI will not be affected by this bug since glade defaults to
GtkBuilder format.