interp/remap: enable remapped G codes as cycles
Assume G84.2 is remapped to Python g842 like so:

def g842(self,**words):
    self.motion_mode = 842 # important
    return INTERP_OK

then executing

    G84.2 x1 y1
    x3 y3
    y5

will execute like:
    G84.2 x1 y1
    G84.2 x3 y3
    G84.2 x3 y5

until motion is cleared with G80 or some other motion is executed.

This enables writing cycles in Python, or as Oword procedures; in the
latter case the self.motion_mode should be set in the Python epilog.