switch port directions in stepconf
This will transpose "In" and "Out" in the combobox for setting port
direction.  I chose this, instead of changing how the values are
treated, because this way users who figured out how to work around the
bug by selecting the "wrong" value will not have their .steconf files
broken when they regenerate them in a future run of stepconf.