new special comment: (AXIS,notify,message)
Problem statement: A gcode subroutine can detect an invalid
user input value and terminate with m2.  If the condition
is detected without generating a toolpath, this abort
may leave the axis gui with no toolpath display and no
informing  messages.

Special comments like:

  (debug, you specified a bad value)

will not be seen unless the run button is pushed but that
is not necessarily obvious or intuitive to a cautious
user when no toolpath is visible.

Special comments like:

   (print,you specified a bad value)

are displayed on stdout but not all users know to look there.

This patch expands the special (AXIS,... comment form to
support a comment like:

  (AXIS,notify,the_message_explaining_the_abort)

Typically, the next line would be an m2 to abort, but it is not mandatory.

Comments lacking a message like:

    (AXIS,notify)
    (AXIS,notify,)

will display the number of uncanceled special comments (they are normally
canceled as displayed so this count will not (currently) accumulate)

An example ngcgui-compatible subfile that demonstrates some behaviors:

o<ntst> sub

        #<zrequest> = #1 (=0)

        (AXIS,notify)            ; just notify
        (AXIS,notify,)           ; nill message, also just notify
        (AXIS,notify,themessage) ; notify with message

o<if0> if [#<zrequest> GT 1]
         (print, zrequest too big #<zrequest> - ABORTING) ;to stdout
         (debug, zrequest too big #<zrequest> - ABORTING) ;not delivered
         (AXIS,notify, zrequest too big - ABORTING)       ;delivered to gui
         m2 ;abort
o<if0> endif
       g0 z #<zrequest>

o<ntst> endsub