interp/introspection: make predefined parameters Python-extensible
if the module 'namedparams' exists (imported from [PYTHON]TOPLEVEL), the names of
callables therein will be become visible as predefined named parameters.

Only names with a leading underscore (_) are added. Existing predefined
named parameters can be redefined; a warning will be generated in this case.

Callables are expected to be functions with 'self' as the first parameter,
'self' being the reference to the interpreter instance. Accepted return types
are int and float; a returned string will set the interpreter error message
and abort execution.

In retrospect, this is how predefined named parameters should have been done to start with -
provide an extension mechanism, and a stock library of predefined
named paramters. However, in this case the stock parameters would have depended on
the Python plugin being configured.