interp/python: always initialize the embedded Python interpreter
This separates instantiation  and configuration of the Python plugin. Also,
always instantiate it - either from task, or the Interp ctor (whoever comes
first). This is safe now since the LD_PRELOAD workaround for GL apps is in
place for all configurations.

there is no point in not enabling the Python plugin it if no TOPLEVEL
script is around. It is now safe to assume that post the Interp() ctor
plugin functions, like accessing internal state, may be
executed. This means that any Python extensions of the interpreter can be
assumed to work regardless if explicitly configured or not.
TOPLEVEL then will only import additional modules.

Conflicts:

	src/emc/rs274ngc/rs274ngc_pre.cc