drop the leading "1:" from debian version numbers
Now that we've renamed our debian package to "linuxcnc", we no longer
need to maintain version number consistency with the old "emc2" packages.

We're going to keep the human-readable part of the version number
consistent with the old scheme, but we're going to drop the "1:" prefix,
it was a work-around for a version numbering bug back at the beginning
of time.