Fix crash in AXIS backplot
The operation on "add_point" has to be a short-circuiting operation or
it dereferences op and oop when they are not valid pointers.

Apparently under optimization some versions of gcc will transform
    add_point |= ...
into a short-circuiting operation, presumably after proving that the RHS
has no side effects.  This, and possibly differing memory layout between
32-bit and 64-bit systems, meant that I hadn't see this bug before Sam
ran into it.

(there's still a modest standards violation here: if npts is 0, then oop
is a pointer to two elements before the start of a block of memory; it
is only permitted to form a pointer up to one element before the start
of a block of memory.  However, in practice no real computer is going to
misbehave just due to this pointer arithmetic)