pncconf -workaround for 7i77_7i76 firmware
this looks for the 7i77_7i76 firmware and converts
the last mux encoder to a reguar encoder (just for pncconf)
this is because of a workaround needed because hostmot2
doesn't allow mux encoders and regular encoders in the same
firmware. HAl in fact will show an eighth encoder but it is
not accessable by the hardware.