pncconf -change spindle enable signal to machine-is-on
if a step gen is used as a spindle drive and the step gen enable
pin is tied to motion-spindle-on then the step gen will
not ramp to zero when S0 or M05 is called. It will stop immediately.
Now only if the machine is turned off or estop is pressed will it try
to stop immediately.