Cleanups stage 1:  linux_* gone; new rt-preempt-user.*, generic rtapi-* files
The various thread systems' code was all copied and pasted, with
changes made to suit that thread system.  The result was a large
amount of duplicated code between all of the thread systems.

This cleanup project endeavors to create generic definitions of the
RTAPI functions declared in rtapi.h.  Thread systems are implemented
by creating a single new .c and .h file for the system.  Standard
macros can be set to override defaults, such as min and max
priorities.  Other macros are used to add extra data to module or task
structs.  And there is a set of macros that tell the RTAPI functions
to call hooks defined in the .c file to effect additional processing.

This commit fully implements the RT_PREEMPT thread system, and does
some work towards the others.  My hopes for the reduction of file
sizes weren't realized, unfortunately, but the entire implementation
of the rt-preempt differences is less than 600 lines.  The structure
is easier to navigate, in my opinion, and there are no more large (or
small) blocks of definition code remaining in the .h files.

- Removed linux_common.h, linux_rtapi.c
- New generic files rtapi_{msg,io,shmem,task,time}.c implement RTAPI functions
  - Removed obsolete vsnprintf.h
- Old threads files will all be replaced
  - Created initial $THREADS.{c,h} files that will replace them
- Build support changes in rtapi/Submakefile, Makefile and configure.in