fold ULAPI in (currently broken); more shuffling
ULAPI has shared code with RTAPI, so it's folded in as well.

Currently broken: the build system needs to build the ULAPI and RTAPI
.o files separately.  There's a mechanism to do this, as can be seen
in the hal code: hal creates a 'hal' directory for ULAPI, and a
'rthal' directoyr for RTAPI so that common sources are compiled into
two places.

(The original RT_PREEMPT patches did this too, but I can't reproduce
it yet.)

Merge of ULAPI:
- rtapi/Submakefile:  Unified ULAPISRCS and RTAPI_APP_SRCS
  - Remove those from preempt-rt
- Put rtapi_init and rtapi_exit into #ifdef block in $(THREADS).c
- Put #ifdef RTAPI blocks around all RTAPI-only code, following rtapi.h
  - Exception is rtapi_io.c; are these used anywhere in ULAPI?

Other shuffling:
- Testing out using mutex_give and mutex_get everywhere
  - This will obviate the need for the pthreads versions in RT_PREEMPT
- Move Xenomai-specific things out of rtapi_common.[ch]
- Move simple_strtol from rtapi_io.c into rtapi_common.c
- Formatting changes