fix a 'lost message' bug in linuxcncrsh
It used to be that if linuxcncrsh did a read(2) and got a partial line
at the end of the buffer, that partial command would get silently dropped.

This commit changes the linuxcncrsh command parsing code to keep the
partial line, and add to it next time it does a read(2).