better interp error message for G10 with bad P-words
It used to say "P value out of range with G10 L1 or G10 L10".  This misses
the possibility that it was a G10 L11, and it confuses the issue by not
telling the user what the actual L-word was.

After this commit it says "P value out of range with G10 L%d", with the
%d replaced with the L-number.

Also fixed the same issue with G10 L2/L20.