gscreen -make it easier to reuse default widgets
If using a slightly modified default screen one would have to duplicate
most of the initializing and updating code. Now u can call separate
functions for the ones you need.
We also add the handlers array to the call to the handlers 'connect_signals'
function so it can in return call gscreens default signal connector.
you would want to do this if you want to add some extra signals after
connecting the default ones.
add a callback to call the funnction that opens and closes the onboard keyboard