fix a t0 bug in Interp::convert_setup_tool
This would accidentally update the tool-related numbered parameters with
the info for T0, when using G10 to update a tool, while no tool is in
the spindle, on random-tc machines.

Most of the time this is probably ok, since by convention T0 is usually
used as the "no tool" too, in which case it has 0 for all its TLOs.
But it's still a bug: #5400 (current tool) would be wrong even in this
case, and the TLO params would be wrong if T0 was used for a real tool.