build system: re-enable hm2_test
The hm2_test module doesn't do anything with PCI or any other kind of IO,
it just fakes reads from a broken FPGA to test that the hostmot2 driver
is robust against wonky firmwares.  So I think it should build just fine
on all the new realtime systems.

This should make the tests pass on the buildbot, on the platforms that
actually build.